+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 
 ## [1.0.0]
 
-- 2.00 ★ check MyBatis SQL
-- 1.00 init
+- 2.01 Copy MyBatis *.xml SQL
+- 2.00 ★ Check MyBatis *.xml SQL
+- 1.00 Export MyBatis *.xml SQL
 
 # 中文更新日志
 
+- 2.01 复制 MyBatis *.xml SQL
 - 2.00 ★ 改用 MyBatis + JSqlParser 检查 SQL 正确性
 - 1.00 使用 Jdom 魔改 + Druid 基本功能